+++ src/main/kotlin/dev/nyon/bbm/config/ModMenuImpl.kt
@@ -1,5 +1,6 @@
 package dev.nyon.bbm.config
 
+/*? if fabric {*/
 import com.terraformersmc.modmenu.api.ConfigScreenFactory
 import com.terraformersmc.modmenu.api.ModMenuApi
 
@@ -11,3 +12,4 @@ class ModMenuImpl : ModMenuApi {
         }
     }
 }
+/*?}*/