+++ changelog.md
@@ -1,1 +1,2 @@
-- add traditional chinese translation
+- add forge and neoforge support
+- change modid from better-boat-movement to bbm